requirements.txt
magic-pdf[full]==1.3.1

fastapi
uvicorn
python-multipart
rq
redis
python-dotenv